Adds generated data.
For various reasons, we might want to undo this - these data are meant
to be generated from the scripts and the raw data. Hence, they likely
should not be committed at all, and rather generated on the fly. This
avoids the chance that these files end up out of sync due to git.

Otherwise, we need to remember to recommit these files each time we
change one of the scripts.